app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile app/requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
altair==5.4.1
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via bs4
blinker==1.8.2
    # via streamlit
bs4==0.0.2
    # via -r app/requirements.in
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
faiss-cpu==1.8.0.post1
    # via -r app/requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   langsmith
    #   openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain==0.2.14
    # via
    #   -r app/requirements.in
    #   langchain-community
langchain-community==0.2.12
    # via
    #   -r app/requirements.in
    #   langchain-experimental
langchain-core==0.2.32
    # via
    #   -r app/requirements.in
    #   langchain
    #   langchain-community
    #   langchain-experimental
    #   langchain-openai
    #   langchain-text-splitters
langchain-experimental==0.0.64
    # via -r app/requirements.in
langchain-openai==0.1.19
    # via -r app/requirements.in
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.117
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.3.0
    # via -r app/requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.22.0
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.6.4
    # via altair
neo4j==5.27.0
    # via -r app/requirements.in
nest-asyncio==1.6.0
    # via -r app/requirements.in
numpy==1.26.4
    # via
    #   faiss-cpu
    #   langchain
    #   langchain-community
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.44.1
    # via
    #   -r app/requirements.in
    #   langchain-openai
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   marshmallow
    #   pytest
    #   streamlit
pandas==2.2.2
    # via
    #   -r app/requirements.in
    #   streamlit
pillow==10.4.0
    # via streamlit
pluggy==1.5.0
    # via pytest
protobuf==5.28.0
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.9.1
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.3
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pypdf==4.3.1
    # via -r app/requirements.in
pytest==8.3.2
    # via -r app/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r app/requirements.in
pytz==2024.1
    # via
    #   neo4j
    #   pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   streamlit
    #   tiktoken
rich==13.8.0
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.34
    # via
    #   langchain
    #   langchain-community
streamlit==1.38.0
    # via -r app/requirements.in
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.7.0
    # via
    #   -r app/requirements.in
    #   langchain-openai
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via pytest
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pypdf
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.2
    # via streamlit
yarl==1.11.1
    # via aiohttp
